+++ NodeSet.h
@@ -4,11 +4,11 @@
 
 typedef struct node_t {
 	uint16_t nodeID;
-	int16_t x;
-	int16_t y;
-	uint16_t r;
-	uint16_t rm;
-	uint16_t c;
+	int32_t x;
+	int32_t y;
+	uint32_t r;
+	uint32_t rm;
+	uint32_t c;
 	bool set;
 	uint8_t setP; // How many intersection points with this node belong to setP
 } node_t;
@@ -33,7 +33,7 @@ public:
 	virtual ~NodeSet();
 
 private:
-	uint16_t distance(int16_t x1, int16_t y1, int16_t x2, int16_t y2);
+	uint16_t distance(int32_t x1, int32_t y1, int32_t x2, int32_t y2);
 	void consolidate();
 	void shell_sort();
